+++ src/tests/test_printer.py
@@ -4,7 +4,7 @@ from tmtccmd.pus_tm.service_1_verification import Service1TmPacked, Service1TM
 from tmtccmd.pus_tm.service_5_event import Service5TM, Service5TmPacked, Severity
 from tmtccmd.pus_tc.service_17_test import pack_service17_ping_command
 from tmtccmd.utility.tmtc_printer import TmTcPrinter, DisplayMode
-from tmtccmd.utility.tmtcc_logger import get_logger, set_tmtc_logger
+from tmtccmd.utility.logger import get_logger, set_tmtc_logger
 
 
 class TestPrinter(TestCase):
