+++ src/tmtccmd/com_if/dummy_com_if.py
@@ -11,7 +11,7 @@ from tmtccmd.pus_tm.factory import PusTelemetryFactory, PusTmListT
 from tmtccmd.pus_tm.service_1_verification import Service1TmPacked
 from tmtccmd.pus_tm.service_17_test import Service17TmPacked
 from tmtccmd.pus.service_17_test import Srv17Subservices
-from tmtccmd.utility.tmtcc_logger import get_logger
+from tmtccmd.utility.logger import get_logger
 
 LOGGER = get_logger()
