+++ src/tmtccmd/com_if/tcpip_udp_com_if.py
@@ -9,7 +9,7 @@ import select
 import socket
 from typing import Union
 
-from tmtccmd.utility.tmtcc_logger import get_logger
+from tmtccmd.utility.logger import get_logger
 from tmtccmd.config.definitions import CoreModeList
 from tmtccmd.com_if.com_interface_base import CommunicationInterface, PusTmListT
 from tmtccmd.pus_tm.factory import PusTelemetryFactory
