+++ src/tmtccmd/defaults/tm_handling.py
@@ -5,7 +5,7 @@
 """
 
 from tmtccmd.ecss.tm import PusTelemetry
-from tmtccmd.utility.tmtcc_logger import get_logger
+from tmtccmd.utility.logger import get_logger
 from tmtccmd.pus_tm.service_1_verification import Service1TM
 from tmtccmd.pus_tm.service_5_event import Service5TM
 from tmtccmd.pus_tm.service_17_test import Service17TM
