+++ src/tmtccmd/pus_tm/service_1_verification.py
@@ -10,7 +10,7 @@ from typing import Deque
 
 from tmtccmd.ecss.tm import PusTelemetry
 from tmtccmd.ecss.tm_creator import PusTelemetryCreator
-from tmtccmd.utility.tmtcc_logger import get_logger
+from tmtccmd.utility.logger import get_logger
 
 LOGGER = get_logger()
