+++ src/tmtccmd/pus_tm/service_8_functional_cmd.py
@@ -4,7 +4,7 @@
 import struct
 
 from tmtccmd.ecss.tm import PusTelemetry
-from tmtccmd.utility.tmtcc_logger import get_logger
+from tmtccmd.utility.logger import get_logger
 
 LOGGER = get_logger()
