+++ src/tmtccmd/utility/exit_handler.py
@@ -1,6 +1,6 @@
 import signal
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
-from tmtccmd.utility.tmtcc_logger import get_logger
+from tmtccmd.utility.logger import get_logger
 
 LOGGER = get_logger()
